# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PhoneMail API",
    "version": "0.1.0",
...
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/api/uploads": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/ExpectedBytes" }, { "$ref": "#/components/parameters/Filename" }, { "$ref": "#/components/parameters/UploadMode" }], "requestBody": { "$ref": "#/components/requestBodies/UploadBytes" }, "responses": { "201": { "description": "Owner-only staged resumable upload or completed bounded stream; mime is signature-derived and scannerState is unscanned" }, "400": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } } } },
    "/api/uploads/{id}/status": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/UploadId" }], "responses": { "200": { "description": "Owner-only resumable upload state for client reconciliation", "content": { "application/json": { "schema": { "type": "object", "required": ["upload"], "properties": { "upload": { "$ref": "#/components/schemas/UploadStatus" } } } } } }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/api/uploads/{id}": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/UploadId" }, { "$ref": "#/components/parameters/Range" }], "responses": { "200": { "description": "Authorized full byte stream" }, "206": { "description": "Authorized single byte range with Content-Range and Accept-Ranges" }, "404": { "$ref": "#/components/responses/Error" }, "416": { "$ref": "#/components/responses/Error" } } }, "patch": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/UploadId" }, { "$ref": "#/components/parameters/UploadOffset" }], "requestBody": { "$ref": "#/components/requestBodies/UploadBytes" }, "responses": { "200": { "description": "Resumable upload committed offset/status" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" } } } },
    "/api/uploads/{id}/attach": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/UploadId" }], "requestBody": { "$ref": "#/components/requestBodies/UploadAttach" }, "responses": { "201": { "description": "Associates an owner-ready upload with an owned message" }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "parameters": {
...
      "UploadId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ExpectedBytes": { "name": "X-Expected-Bytes", "in": "header", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 10485760 } },
      "Filename": { "name": "X-Filename", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 120 } },
      "UploadMode": { "name": "X-Upload-Mode", "in": "header", "required": false, "schema": { "type": "string", "enum": ["resumable"] } },
      "UploadOffset": { "name": "X-Upload-Offset", "in": "header", "required": true, "schema": { "type": "integer", "minimum": 0 } },
      "Range": { "name": "Range", "in": "header", "required": false, "schema": { "type": "string", "pattern": "^bytes=[0-9]*-[0-9]*$" } },
...
    },
    "requestBodies": {
...
      "UploadBytes": { "required": true, "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
      "UploadAttach": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["messageId"], "additionalProperties": false, "properties": { "messageId": { "type": "string", "format": "uuid" } } } } } },
...
    },
    "schemas": {
...
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "requestId", "retryable"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" }, "fields": { "type": "object", "additionalProperties": true } } } } },
...
      "UploadStatus": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "expectedBytes": { "type": "integer" }, "offset": { "type": "integer" }, "status": { "type": "string", "enum": ["staged", "ready", "expired"] }, "scannerState": { "type": "string", "enum": ["unscanned", "pending", "clean", "rejected"] }, "expiresAt": { "type": "string", "format": "date-time" } } },
...
    },
    "responses": {
...
      "Error": { "description": "Stable error envelope with code, message, requestId and retryable", "headers": { "X-Request-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" }, "description": "Present for rate limits and retryable transient failures" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_session" },
      "refreshCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_refresh" }
    }
  }
}